+++ CHANGELOG.md
@@ -34,6 +34,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - credits modal: added new section for translations
 - custom types: expose a switch/outlet as a washing machine
 - refreshed icons for fan, television, speaker + filter
+- added confirmation modal for shutting down
 
 ### Other Changes